pyproject.toml
[build-system]
requires = ["setuptools==75.1.0", "wheel==0.44.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bootstrappedlm"
version = "0.1.0"
description = "Skeleton package for a Zipfian world-model agent"
readme = "Outline.md"
requires-python = ">=3.11"
dependencies = [
  "numpy==2.1.3",
  "PyYAML==6.0.2",
  "torch==2.5.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
